User Authorization to edit and delete

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,10 @@ class ListingController extends Controller
 
     // Update Listing data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action!');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -78,6 +82,11 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action!');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
